+++ Source/Processor/ProcessorLevelData.cpp
@@ -20,12 +20,12 @@ ProcessorLevelData::~ProcessorLevelData()
 
 }
 
-void ProcessorLevelData::SetLevel(int channel, ProcessorLevelData::LevelVal level)
+void ProcessorLevelData::SetLevel(unsigned long channel, ProcessorLevelData::LevelVal level)
 {
     m_levelMap[channel] = level;
 }
 
-ProcessorLevelData::LevelVal ProcessorLevelData::GetLevel(int channel)
+ProcessorLevelData::LevelVal ProcessorLevelData::GetLevel(unsigned long channel)
 {
     if(m_levelMap.count(channel))
         return m_levelMap.at(channel);
@@ -33,12 +33,12 @@ ProcessorLevelData::LevelVal ProcessorLevelData::GetLevel(int channel)
         return ProcessorLevelData::LevelVal(0,0);
 }
 
-void ProcessorLevelData::SetChannelCount(int count)
+void ProcessorLevelData::SetChannelCount(unsigned long count)
 {
     if(GetChannelCount()==count)
         return;
     
-    for(int i=1; i<=count; i++)
+    for(unsigned long i=1; i<=count; i++)
     {
         if(!m_levelMap.count(i))
             m_levelMap[i] = LevelVal(0,0);
